requirements.txt
tushare
python-dotenv
numpy
pandas
